Refactor coreaudio host and add iOS support + ios-feedback example
