+++ examples/ios-feedback/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cpal-ios-example"
+version = "0.1.0"
+authors = ["Michael Hills <mhills@gmail.com>"]
+edition = "2018"
+
+[lib]
+name = "cpal_ios_example"
+crate-type = ["staticlib"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+cpal = { path = "../.." }
+anyhow = "1.0.12"
+ringbuf = "0.1.6"
+
