+++ examples/ios-feedback/ios-src/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  cpal-ios-example
+//
+//  Created by Michael Hills on 2/10/20.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@end
+
