+++ examples/ios-feedback/ios-src/main.m
@@ -1,0 +1,18 @@
+//
+//  main.m
+//  cpal-ios-example
+//
+//  Created by Michael Hills on 2/10/20.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    NSString * appDelegateClassName;
+    @autoreleasepool {
+        // Setup code that might create autoreleased objects goes here.
+        appDelegateClassName = NSStringFromClass([AppDelegate class]);
+    }
+    return UIApplicationMain(argc, argv, nil, appDelegateClassName);
+}
