+++ examples/ios-feedback/src/lib.rs
@@ -1,0 +1,6 @@
+mod feedback;
+
+#[no_mangle]
+pub extern "C" fn rust_ios_main() {
+    feedback::run_example().unwrap();
+}
